refactor: create widget
