+++ src/app/(pages)/profile/[id]/page.tsx
@@ -1,8 +1,8 @@
 import { UserProfile } from '@/features/profile/ui/userProfile'
 import { GetPostsResponse } from '@/features/post/api/postsApi.types'
-import { PostsGridWithInfiniteScroll } from '@/features/post/ui/postsGridWithInfiniteScroll/PostsGridWithInfiniteScroll'
 import { PostModal } from '@/features/post/ui/postModal'
 import { PublicUserProfile } from '@/features/profile/api/profileApi.types'
+import { PostsGridWithInfiniteScroll } from '@/features/post/ui/postsGridWithInfiniteScroll'
 
 type PageProps = {
   params: Promise<{ id: string }>
