+++ src/features/post/ui/postsGridWithInfiniteScroll/index.ts
@@ -1,0 +1,1 @@
+export * from './PostsGridWithInfiniteScroll'
