+++ src/shared/ui/index.ts
@@ -24,7 +24,7 @@ export { Recaptcha } from './reCaptcha'
 export { Scroll } from './scroll'
 export { SelectBox, SelectItem } from './selectBox'
 export { Separator } from './separator'
-export { Sidebar, createSidebarItems } from './sidebar'
+export { Sidebar, createSidebarItems } from '../../widgets/sidebar'
 export { Tabs } from './tabs'
 export { TextArea } from './textArea'
 export { TextField } from './textField'
