+++ src/widgets/header/Header.tsx
@@ -6,10 +6,10 @@ import Image, { StaticImageData } from 'next/image'
 import { IoNotificationsOutline } from 'react-icons/io5'
 import { Button, SelectBox, SelectItem } from '@/shared/ui'
 import { PATH } from '@/shared/config/routes'
-import Letters from '../../../../public/logo/letters.png'
-import Logo from '../../../../public/logo/logo.png'
-import flagUnitedKingdom from '@/shared/ui/header/assets/flagUnitedKingdom.png'
-import flagRussia from '@/shared/ui/header/assets/flagRussia.png'
+import Letters from '../../../public/logo/letters.png'
+import Logo from '../../../public/logo/logo.png'
+import flagUnitedKingdom from './assets/flagUnitedKingdom.png'
+import flagRussia from './assets/flagRussia.png'
 import s from './Header.module.scss'
 import { useGetMeQuery } from '@/features/auth/api/authApi'
