+++ src/widgets/header/assets/flagRussia.png
