+++ src/widgets/sidebar/Sidebar.module.scss
