+++ src/widgets/sidebar/SidebarItem.tsx
